+++ src/exceptions/TestPreflightException.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Notifier plugin for Craft CMS
+ *
+ * First-class Notifications for Craft CMS
+ *
+ * @author    Double Secret Agency
+ * @link      https://plugins.doublesecretagency.com/
+ * @copyright Copyright (c) 2021 Double Secret Agency
+ */
+
+namespace doublesecretagency\notifier\exceptions;
+
+use Exception;
+
+/**
+ * Class TestPreflightException
+ * @since 3.0.0
+ */
+class TestPreflightException extends Exception
+{
+}